+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 
 ### Changed
 
-- `Sepiphy\Config\Config` loads both php and yml files of a directory by default, no need to care about loader instance.
+- `Sepiphy\Config\Config` loads both php and yml files of a directory by default, no need to care about loaders anymore.